Add information for slashing into `BlockHeader`
As a preparation for slashing bad validators, we need inforamtion about that
and need to store the information for consensus. So we add two important
information, preimage_root and missing_preimage_validators. The `preimage_root`
is the hash of the merkle root of the preimages for the height.
The `missing_preimage_validators` is a list of indices to the validator UTXO
set which hae not revealed the preimage.