Correct the formula for calculating height of preimage
We calculate the height of preimage when getting a pre-image. Simply, the preimage
for the next height of an enrolled height must be the preimage of distance 0. But
the old formula get the preimage of one more higher distance like 1 in this case.
So we correct this and modify test case accordingly.