Allow setting quorums when nomination stops
On a block being externalized, a current nomination round stops,
and the newly constructed quorums are set. But there is a situation
where the previously called nomination process is running while
setting new quorums, which is a probable situation but is not
accounted for in our code. So we allow setting new quorums even
while a previous called nomination is running if the nomination is
for the next height.